f1tenth/requirements-lock.txt
annotated-types==0.7.0
asttokens==3.0.1
casadi==3.7.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
clarabel==0.11.1
click==8.3.1
cloudpickle==3.1.2
comm==0.2.3
contourpy==1.3.3
cvxpy==1.8.1
cvxpylayers==1.0.1
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.20
decorator==5.2.1
diffcp==1.1.8
executing==2.2.1
f1tenth-gym @ git+https://github.com/f1tenth/f1tenth_gym.git@67bc6dbe5826e3160fa97116d4efbb51b7a9670f
-e file:///home/vietanhle/github/UPenn_Py/L2O_MPPI/f1tenth_planning
farama-notifications==0.0.4
filelock==3.20.0
fonttools==4.61.1
fsspec==2025.12.0
gitdb==4.0.12
gitpython==3.1.46
gurobipy==13.0.1
gymnasium==0.29.1
highspy==1.13.1
idna==3.11
ipykernel==7.2.0
ipython==9.10.0
ipython-pygments-lexers==1.1.1
jax==0.4.33
jax-cuda12-pjrt==0.4.33
jax-cuda12-plugin==0.4.33
jaxlib==0.4.33
jedi==0.19.2
jinja2==3.1.6
joblib==1.5.3
jupyter-client==8.8.0
jupyter-core==5.9.1
kiwisolver==1.4.9
llvmlite==0.46.0
markupsafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
ml-dtypes==0.5.4
mpmath==1.3.0
mypy-extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6.1
numba==0.64.0
numpy==2.4.2
nvidia-cublas-cu12==12.9.1.4
nvidia-cuda-cupti-cu12==12.9.79
nvidia-cuda-nvcc-cu12==12.9.86
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.9.79
nvidia-cudnn-cu12==9.19.0.56
nvidia-cufft-cu12==11.4.1.4
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.7.5.82
nvidia-cusparse-cu12==12.5.10.65
nvidia-nccl-cu12==2.29.3
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
opencv-python==4.13.0.92
opt-einsum==3.4.0
osqp==1.1.1
packaging==26.0
pandas==2.3.3
parso==0.8.6
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.9.2
prompt-toolkit==3.0.52
protobuf==6.33.5
psutil==7.2.2
ptyprocess==0.7.0
pure-eval==0.2.3
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pygame==2.6.1
pygments==2.19.2
pyparsing==3.3.2
pyqt6==6.10.2
pyqt6-qt6==6.10.2
pyqt6-sip==13.11.0
pyqtgraph==0.13.7
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.3
pyzmq==27.1.0
qdldl==0.1.9.post1
requests==2.32.5
scipy==1.17.1
scs==3.2.11
sentry-sdk==2.53.0
setuptools==82.0.0
shapely==2.1.2
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
sympy==1.13.1
threadpoolctl==3.6.0
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
triton==3.1.0
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
tzdata==2025.3
urllib3==2.6.3
wandb==0.25.0
wcwidth==0.6.0
yamldataclassconfig==1.5.0
